EmbeddingService 리팩토링 및 Vertex AI 연동 : feat : AI의존성 common로직>AI, MEMBER<AI의존, ITEM<AI의존 https://github.com/TEAM-ROMROM/RomRom-BE/issues/229
