+++ RomRom-Domain-AI/build.gradle
@@ -13,7 +13,6 @@ jar {
 
 dependencies {
     api project(':RomRom-Common')
-    api project(':RomRom-Domain-Item')
     
     // Google OAuth2 & Jackson & OkHttp for Vertex AI
     implementation 'com.google.auth:google-auth-library-oauth2-http:1.19.0'
