+++ RomRom-Domain-AI/src/main/java/com/romrom/ai/service/VertexAiClient.java
@@ -1,4 +1,4 @@
-package com.romrom.ai;
+package com.romrom.ai.service;
 
 import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.JsonNode;
