+++ RomRom-Domain-Member/build.gradle
@@ -13,6 +13,7 @@ jar {
 
 dependencies {
     api project(':RomRom-Common')
+    api project(':RomRom-Domain-AI')      // AI 기능(임베딩, 가격 예측) 사용
     
     // 도메인 로직에만 필요한 추가 의존성이 있다면 여기에 추가
 }
